before add dump function

+++ GaussLegendre.m
@@ -3,9 +3,9 @@
 
 function [ x, w ] = GaussLegendre(m)
 
-% m should be int32
+% m should be int32 to GaussLegendreMex
 
-[ x, w ] = GaussLegendreMex(m);
+[ x, w ] = GaussLegendreMex(int32(m));
 
 return
